Renamed component from jsx to tsx
