+++ src/components/SeasonalGraphic.tsx
